src/ur10_ft300_monitor/requirements.txt
numpy
pandas
pyqt5
pyqtgraph
pyserial
